Refactor to split component and use hoisted state
